Implement settings with SwiftUI/Combine
This updates the settings TableViewController to a SwiftUI view in a
UIHostingController. I was interested in seeing what it would look like
as SwiftUI code and whether it would be simpler to create. Not sure I
have a definitive answer to that, some things were better, some were
not. But it does seem easier to iterate on as a SwiftUI view.

In making the change to SwiftUI I also updated the Settings object to be
an ObservableObject. This change was straightforward and works well with
the new SwiftUI view (as should be expected).